cogni-knowledge 0.0.18: extract _knowledge_lib.py (closes #272, #273)
Two duplicated copies of `normalize_url` + `_STRIP_QUERY_*` + the
`tempfile.mkstemp + os.replace` atomic-write pattern lived in
`candidate-store.py` and `fetch-cache.py`. The `normalize_url` bodies had
already started style-drifting between the two callers (one binds
`query = urlencode(kept)`, the other inlines it). Identical output today,
but the dedup-key contract — curator-side merge into `candidates.json`
must agree byte-for-byte with the fetch-cache lookup key — depends on
that identity.

Extract into `scripts/_knowledge_lib.py` as the shared source of truth.
The two existing scripts now import from there; `_atomic_write` call
sites renamed in-place to the public `atomic_write`. No behavioural
change. New `tests/test_knowledge_lib.sh` asserts three-way `is`-identity
of the shared symbols across callers (structural rather than
convention), behavioural canonicalization across a representative URL,
and an `atomic_write` round-trip with no `.tmp` debris.

The extraction was originally scheduled for M5 alongside `source-ingester`
as the third caller; landed early because the style drift is the kind of
thing that becomes a real bug right before M5 ships.

CHANGELOG also rewrites the v0.0.17 sentence about the `0.0.16` slot to
reference the new annotated tag `cogni-knowledge-v0.0.16-alpha-measurement`
at commit f6c9d24e (issue #273 acceptance item — the tag itself is
pushed separately).

`knowledge-binding.py:_write_binding` shares the same atomic-write
pattern but with a different signature (takes `knowledge_root`, resolves
the binding path internally). Left untouched this round and noted in
CHANGELOG as a possible follow-up.
