+++ cogni-knowledge/scripts/_knowledge_lib.py
@@ -1,0 +1,93 @@
+#!/usr/bin/env python3
+"""
+_knowledge_lib.py — shared primitives for cogni-knowledge scripts.
+
+Single source of truth for URL identity in the v0.1.0 inverted pipeline.
+`candidate-store.py` (curator-side merge into `candidates.json`) and
+`fetch-cache.py` (fetcher-side cache lookup keyed by `sha256(normalize_url(url))`)
+must agree byte-for-byte on the canonical form of any URL — otherwise the
+curator-side dedup and the fetch-cache hit/miss decision drift, and a URL
+present in `candidates.json` can silently miss in the cache (or vice versa).
+
+Before this module existed, `normalize_url` + the `_STRIP_QUERY_*` constants
++ the `tempfile.mkstemp + os.replace` atomic-write pattern were duplicated
+across both scripts. The copies had already started style-drifting on
+`normalize_url`. The deferred extraction was originally scheduled for M5
+(`source-ingester` as a third caller); landed early at v0.0.18 to close
+the divergence before it became behavioural.
+
+`knowledge-binding.py:_write_binding` shares the same atomic-write pattern
+but a different signature (takes `knowledge_root`, resolves the binding
+path internally). Not extracted here — possible follow-up.
+
+stdlib-only. Python 3.8+.
+"""
+
+from __future__ import annotations
+
+import json
+import os
+import tempfile
+from pathlib import Path
+from urllib.parse import urlsplit, urlunsplit, parse_qsl, urlencode
+
+# Tracking-param prefixes/names stripped during URL normalization. Covers
+# the common cases seen in EU regulatory crawls; conservative on purpose —
+# only well-known tracking params are dropped to avoid breaking URLs that
+# rely on a query string for content identity.
+_STRIP_QUERY_PREFIXES = ("utm_",)
+_STRIP_QUERY_EXACT = frozenset({"ref", "fbclid", "gclid"})
+
+
+def normalize_url(url: str) -> str:
+    """Canonicalize a URL for dedup purposes.
+
+    - Lowercase scheme + host (path case preserved — RFC 3986 §6.2.2.1).
+    - Strip trailing `/` from path unless the path is just `/`.
+    - Drop query params whose name starts with any _STRIP_QUERY_PREFIXES
+      or is in _STRIP_QUERY_EXACT. Remaining params keep their order.
+    - Drop fragment.
+
+    Whitespace-only or empty input returns the input unchanged so callers
+    surface the bad value rather than silently coalescing distinct entries.
+    """
+    if not url or not url.strip():
+        return url
+    parts = urlsplit(url.strip())
+    scheme = parts.scheme.lower()
+    netloc = parts.netloc.lower()
+    path = parts.path
+    if path.endswith("/") and path != "/":
+        path = path.rstrip("/")
+    kept = [
+        (k, v)
+        for k, v in parse_qsl(parts.query, keep_blank_values=True)
+        if not (k.startswith(_STRIP_QUERY_PREFIXES) or k in _STRIP_QUERY_EXACT)
+    ]
+    query = urlencode(kept)
+    return urlunsplit((scheme, netloc, path, query, ""))
+
+
+def atomic_write(path: Path, payload: dict) -> Path:
+    """Atomically write `payload` as pretty-printed JSON to `path`.
+
+    `tempfile.mkstemp` so two concurrent writers to the same target cannot
+    collide on a predictable `.tmp` suffix; `os.replace` for the atomic
+    swap; tmp-file unlink on exception so failures don't leave debris.
+    The temp file lives in the same directory as the target so the
+    `os.replace` cannot cross devices.
+    """
+    path.parent.mkdir(parents=True, exist_ok=True)
+    fd, tmp = tempfile.mkstemp(prefix=f".{path.name}.", suffix=".tmp", dir=str(path.parent))
+    try:
+        with os.fdopen(fd, "w", encoding="utf-8") as fh:
+            json.dump(payload, fh, indent=2, ensure_ascii=False)
+            fh.write("\n")
+        os.replace(tmp, path)
+    except Exception:
+        try:
+            os.unlink(tmp)
+        except OSError:
+            pass
+        raise
+    return path
